+++ docs/changelog.md
@@ -12,6 +12,7 @@
 - [X] Adicionado o MFA para o login
 - [X] Adicionado a role global para os tokens de microservices
 - [X] Adicionado a troca de email e update do user
+- [X] Adicionado o soft delete do usuário
 
 # 0.1.0
 - [X] Adicionado o arquivo de changelog